CookieBanner Progress - CookieMap added and banner functionality fixed
