+++ src/explore-education-statistics-frontend/src/components/CookieBanner.tsx
@@ -1,42 +1,73 @@
 import React from 'react';
+import { NextContext } from 'next';
+import cookie from 'cookie';
 import { useCookies } from 'react-cookie';
 import ButtonText from '@common/components/ButtonText';
-import * as googleAnalyticsService from '@frontend/services/googleAnalyticsService';
-import {
-  acceptCookies,
-  getCookies,
-  bannerSeenCookieName,
-  cookieSettingsCookieName,
-} from '@frontend/services/cookiesService';
+import CookieMap from '@frontend/services/cookieMap';
 import useMounted from 'explore-education-statistics-common/src/hooks/useMounted';
 import styles from './CookieBanner.module.scss';
 
-const CookieBanner = () => {
-  useMounted(() => {
-    if (getCookies()[cookieSettingsCookieName] !== false) {
-      googleAnalyticsService.initGA();
-    }
-  });
-  const [cookies, setCookie] = useCookies([bannerSeenCookieName]);
+interface Props {
+  cookies?: any;
+}
 
-  return cookies[bannerSeenCookieName] !== 'true' ? (
-    <div className={styles.container}>
-      <p>
-        <span>GOV.UK uses cookies to make the site simpler.</span>{' '}
-        <ButtonText
-          type="button"
-          className={styles.button}
-          onClick={() => {
-            acceptCookies();
-          }}
-        >
-          Accept Cookies
-        </ButtonText>{' '}
-        or{' '}
-        <a href="/cookies">find out more about cookies and cookie settings</a>.
-      </p>
-    </div>
-  ) : null;
+function CookieBanner({ cookies }: Props) {
+  const [liveCookies, setCookie] = useCookies();
+
+  const { isMounted } = useMounted();
+
+  const acceptCookies = () => {
+    setCookie(CookieMap.bannerSeenCookie.name, true, {
+      expires: CookieMap.bannerSeenCookie.expires,
+    });
+    if (liveCookies[CookieMap.disableGACookie.name] === undefined) {
+      setCookie(CookieMap.disableGACookie.name, false, {
+        expires: CookieMap.disableGACookie.expires,
+      });
+    }
+  };
+
+  function render() {
+    return (
+      <div className={styles.container}>
+        <p>
+          <span>GOV.UK uses cookies to make the site simpler.</span>{' '}
+          <ButtonText
+            type="button"
+            className={styles.button}
+            onClick={() => {
+              acceptCookies();
+            }}
+          >
+            Accept Cookies
+          </ButtonText>{' '}
+          or{' '}
+          <a href="/cookies">find out more about cookies and cookie settings</a>
+          .
+        </p>
+      </div>
+    );
+  }
+
+  if (isMounted) {
+    console.log('mounted render', liveCookies);
+    return liveCookies[CookieMap.bannerSeenCookie.name] === 'true'
+      ? null
+      : render();
+  }
+  console.log('unmounted render', cookies);
+  return cookies[CookieMap.bannerSeenCookie.name] === 'true' ? null : render();
+}
+
+CookieBanner.getInitialProps = (props: NextContext) => {
+  if (props && props.req && props.req.headers && props.req.headers.cookie) {
+    return {
+      cookies: cookie.parse(props.req.headers.cookie),
+    };
+  }
+  return {
+    cookies: {},
+  };
 };
 
 export default CookieBanner;
